refactor: 객체 타입 안정성 강화, 오류 코드의 유일성 확보, 로깅 수준 조정
