--- src/main/java/org/creditto/credittoService/global/response/error/ErrorCode.java
@@ -1,6 +1,0 @@
-package org.creditto.credittoService.global.response.error;
-
-import org.creditto.credittoService.global.response.ApiCode;
-
-public interface ErrorCode extends ApiCode {
-}
