+++ src/main/java/org/creditto/creditto_service/global/common/BaseEntity.java
@@ -1,4 +1,4 @@
-package org.creditto.credittoService.global.common;
+package org.creditto.creditto_service.global.common;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
