+++ src/main/java/org/creditto/creditto_service/global/response/ApiCode.java
@@ -1,4 +1,4 @@
-package org.creditto.credittoService.global.response;
+package org.creditto.creditto_service.global.response;
 
 import org.springframework.http.HttpStatus;
