+++ src/main/java/org/creditto/creditto_service/global/response/error/ErrorBaseCode.java
@@ -1,4 +1,4 @@
-package org.creditto.credittoService.global.response.error;
+package org.creditto.creditto_service.global.response.error;
 
 import lombok.AccessLevel;
 import lombok.Getter;
