+++ src/main/java/org/creditto/creditto_service/global/response/error/ErrorCode.java
@@ -1,0 +1,6 @@
+package org.creditto.creditto_service.global.response.error;
+
+import org.creditto.creditto_service.global.response.ApiCode;
+
+public interface ErrorCode extends ApiCode {
+}
